fix: resolve compilation errors and remove legacy V1 TAP code
- Fix BigDecimal type inference in config validation and GRT parsing
- Collapse else-if in validation.rs per clippy suggestion
- Remove legacy TapReceipt::V1 handling and tests (V1/Legacy removed)
- Fix undefined variable typos in tests (allocation_address, mock_network_subgraph)
- Add AllocationId import for CollectionId conversion
- Allow deprecated bearer auth warning (TODO for future cleanup)

+++ integration-tests/src/database_checker.rs
@@ -811,7 +811,7 @@ impl DatabaseChecker {
                     .unwrap_or(0.0)
             );
 
-            if total_pending_value >= BigDecimal::from(trigger_value) {
+            if total_pending_value >= trigger_value {
                 println!("   ✅ Ready to trigger RAV!");
             } else {
                 let needed = BigDecimal::from(trigger_value) - &total_pending_value;
